requirements.txt
numpy>=1.20.0
torch>=1.8.0
matplotlib>=3.4.0
tqdm>=4.60.0
pyyaml>=5.4.0
seaborn>=0.11.0
pandas>=1.2.0
scipy>=1.6.0 